% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{add_item2}
\alias{add_item2}
\title{Define parameters that may be used in model calculations (uses expressions)}
\usage{
add_item2(.data = NULL, input)
}
\arguments{
\item{.data}{Existing data}

\item{input}{Items to define for the simulation as an expression (i.e., using {})}
}
\value{
A substituted expression to be evaluated by engine
}
\description{
Define parameters that may be used in model calculations (uses expressions)
}
\details{
DEPRECATED (old description): The functions to add/modify events/inputs use named vectors or lists. If chaining together add_item2, it will just append the expressions together in the order established.

If using \code{pick_val_v}, note it should be used with the \code{deploy_env = TRUE} argument so that add_item2 process it correctly.
}
