% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{add_other}
\alias{add_other}
\title{Defining other data for events and interventions (only ongoing and instantaneous)}
\usage{
add_other(.data = NULL, other, evt, arm, category = "default")
}
\arguments{
\item{.data}{Existing other data}

\item{other}{Value or expression to calculate other estimates}

\item{evt}{Events for which this other data is applicable}

\item{arm}{Interventions for which this other data is applicable}

\item{category}{Name of the other data category. If not defined, will be set as default.}
}
\value{
A list of other data
}
\description{
Defining other data for events and interventions (only ongoing and instantaneous)
}
\details{
Other data can be defined by writing expressions and objects in the value argument whose execution will be delayed until the model runs.

This function accepts the use of pipes (\%>\%) to define multiple other data.
}
\examples{
\dontrun{
add_other(evt = c("start","idfs","ttot"),
arm = c("int", "noint"),
other = 1,
category = "other_discounted_output"
)
}
}
