% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{conditional_qlnorm}
\alias{conditional_qlnorm}
\title{Conditional quantile function for lognormal distribution}
\usage{
conditional_qlnorm(rnd = 0.5, meanlog, sdlog, lower_bound)
}
\arguments{
\item{rnd}{Vector of quantiles}

\item{meanlog}{The meanlog parameter}

\item{sdlog}{The sdlog parameter}

\item{lower_bound}{The lower bound to be used (current time)}
}
\value{
Estimate(s) from the conditional lognormal distribution based on given parameters
}
\description{
Conditional quantile function for lognormal distribution
}
\examples{
conditional_qlnorm(rnd = 0.5, meanlog = 1,sdlog = 1,lower_bound = 1)
}
