% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{conditional_qweibull}
\alias{conditional_qweibull}
\title{Conditional quantile function for weibull distribution}
\usage{
conditional_qweibull(rnd = 0.5, shape, scale, lower_bound)
}
\arguments{
\item{rnd}{Vector of quantiles}

\item{shape}{The shape parameter}

\item{scale}{The scale parameter}

\item{lower_bound}{The lower bound to be used (current time)}
}
\value{
Estimate(s) from the conditional weibull distribution based on given parameters
}
\description{
Conditional quantile function for weibull distribution
}
\examples{
conditional_qweibull(rnd = 0.5,shape = 1,scale = 1,lower_bound = 1)
}
