% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{create_indicators}
\alias{create_indicators}
\title{Creates a vector of indicators (0 and 1) for sensitivity/DSA analysis}
\usage{
create_indicators(sens, n_sensitivity, n_elem, n_elem_before = 0)
}
\arguments{
\item{sens}{current analysis iterator}

\item{n_sensitivity}{total number of analyses to be run}

\item{n_elem}{number of elements to iterate through}

\item{n_elem_before}{number of elements that go before the current n_elem to iterate through}
}
\value{
Numeric vector composed of 0 and 1, where value 1 will be used by \code{pick_val_v} to pick the corresponding index in its \code{sens} argument
}
\description{
Creates a vector of indicators (0 and 1) for sensitivity/DSA analysis
}
\details{
n_elem_before is to be used when several indicators want to be used (e.g., for patient level and common level inputs) while facilitating readibility of the code
}
\examples{
\dontrun{
create_indicators(10,20,4)
create_indicators(11,20,10,2)
}
}
