% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{disc_instant_v}
\alias{disc_instant_v}
\title{Calculate instantaneous discounted costs or qalys for vectors}
\usage{
disc_instant_v(lcldr = 0.035, lclcurtime, lclval)
}
\arguments{
\item{lcldr}{The discount rate}

\item{lclcurtime}{The time of the current event in the simulation}

\item{lclval}{The value to be discounted}
}
\value{
Double based on discrete time discounting
}
\description{
Calculate instantaneous discounted costs or qalys for vectors
}
\examples{
\dontrun{
disc_instant_v(lcldr=0.035, lclcurtime=3, lclval=2500)
}

}
