% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{disc_ongoing}
\alias{disc_ongoing}
\title{Calculate discounted costs and qalys between events}
\usage{
disc_ongoing(lcldr = 0.035, lclprvtime, lclcurtime, lclval)
}
\arguments{
\item{lcldr}{The discount rate}

\item{lclprvtime}{The time of the previous event in the simulation}

\item{lclcurtime}{The time of the current event in the simulation}

\item{lclval}{The value to be discounted}
}
\value{
Double based on continuous time discounting
}
\description{
Calculate discounted costs and qalys between events
}
\examples{
\dontrun{
disc_ongoing(lcldr=0.035,lclprvtime=0.5, lclcurtime=3, lclval=2500)
}

}
