% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{discrete_resource_clone}
\alias{discrete_resource_clone}
\title{Clone independent discrete resources}
\usage{
discrete_resource_clone(x, n = 1)
}
\arguments{
\item{x}{discrete resource created with resource_discrete()}

\item{n}{Number of independent clones to be generated}
}
\value{
List of independent clones of discrete resource envs (even for n = 1)
}
\description{
Clone independent discrete resources
}
