% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{.resource_discrete_from_xptr}
\alias{.resource_discrete_from_xptr}
\title{Creates a discrete resource from an external pointer}
\usage{
.resource_discrete_from_xptr(xptr)
}
\arguments{
\item{xptr}{external pointer from resource_discrete()$.ptr}
}
\value{
new cloned discrete resource
}
\description{
Creates a discrete resource from an external pointer
}
\keyword{internal}
