% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{get_event}
\alias{get_event}
\title{Get a specific event time}
\usage{
get_event(event_name, ptr, patient_id)
}
\arguments{
\item{event_name}{Character string, the name of the event.}

\item{ptr}{The event queue pointer. Defaults to \code{cur_evtlist}.}

\item{patient_id}{The patient ID. Defaults to \code{i}.}
}
\value{
Numeric, time of event for patient
}
\description{
Get a specific event time
}
