% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{pick_val}
\alias{pick_val}
\title{Select which value should be applied in the corresponding loop}
\usage{
pick_val(
  base,
  psa,
  sens,
  psa_ind = psa_bool,
  sens_ind = sens_bool,
  indicator = 0
)
}
\arguments{
\item{base}{Value if no PSA/DSA/Scenario}

\item{psa}{Value if PSA}

\item{sens}{Value if sensitivity (DSA/Scenario)}

\item{psa_ind}{Boolean whether PSA is active}

\item{sens_ind}{Boolean whether Scenario/DSA is active}

\item{indicator}{Indicator which checks whether the specific parameter is active in the DSA or Scenario loop}
}
\value{
Value used for the inputs
}
\description{
Select which value should be applied in the corresponding loop
}
\examples{
\dontrun{
pick_val(base = 0, psa =rnorm(1,0,0.1), sens = 5,psa_ind = TRUE, sens_ind = FALSE, indicator=1)
}

}
