% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{pop_event}
\alias{pop_event}
\title{Remove the next event from the queue}
\usage{
pop_event(ptr)
}
\arguments{
\item{ptr}{The event queue pointer. Defaults to \code{cur_evtlist}.}
}
\value{
NULL (invisible). Modifies the queue in-place.
}
\description{
Removes the next scheduled event from the queue. Not needed by user.
}
