% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{presgompertz}
\alias{presgompertz}
\title{Survival Probaility function for restricted Gompertz distribution (lower bound only)}
\usage{
presgompertz(time = 1, shape, rate, lower_bound)
}
\arguments{
\item{time}{Vector of times}

\item{shape}{The shape parameter of the Gompertz distribution, defined as in the coef() output on a flexsurvreg object}

\item{rate}{The rate parameter of the Gompertz distribution, defined as in the coef() output on a flexsurvreg object}

\item{lower_bound}{The lower bound of the restricted distribution}
}
\value{
Estimate(s) from the restricted Gompertz distribution based on given parameters
}
\description{
Survival Probaility function for restricted Gompertz distribution (lower bound only)
}
\examples{
presgompertz(time=1,shape=0.05,rate=0.01,lower_bound = 50)
}
