% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{print.resource_discrete}
\alias{print.resource_discrete}
\title{Print method for resource_discrete}
\usage{
\method{print}{resource_discrete}(x, ...)
}
\arguments{
\item{x}{A resource_discrete object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for resource_discrete
}
\keyword{internal}
