% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary_f.R
\name{print.warden_results}
\alias{print.warden_results}
\title{Print method for warden_results}
\usage{
\method{print}{warden_results}(x, ...)
}
\arguments{
\item{x}{A \code{warden_results} object returned by \code{\link[=run_sim]{run_sim()}} or
\code{\link[=run_sim_parallel]{run_sim_parallel()}}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a summary of the structure of a WARDEN simulation results object.
}
