% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{qbeta_mse}
\alias{qbeta_mse}
\title{Draw from a beta distribution based on mean and se (quantile)}
\usage{
qbeta_mse(q, mean_v, se)
}
\arguments{
\item{q}{Quantiles to be used}

\item{mean_v}{A vector of the mean values}

\item{se}{A vector of the standard errors of the means}
}
\value{
A single estimate from the beta distribution based on given parameters
}
\description{
Draw from a beta distribution based on mean and se (quantile)
}
\examples{
qbeta_mse(q=0.5,mean_v=0.8,se=0.2)
}
