% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{qcategorical}
\alias{qcategorical}
\title{Quantile function for categorical distribution}
\usage{
qcategorical(p, probs)
}
\arguments{
\item{p}{Numeric between 0 and 1 (inclusive). Probability (same as \code{luck}
in \code{\link[=draw_categorical]{draw_categorical()}}).}

\item{probs}{Numeric vector of probabilities.}
}
\value{
Integer: 1-based index.
}
\description{
Alias for \code{\link[=draw_categorical]{draw_categorical()}} using standard quantile-function naming.
}
