% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{qcond_exp}
\alias{qcond_exp}
\title{Conditional quantile function for exponential distribution}
\usage{
qcond_exp(rnd = 0.5, rate)
}
\arguments{
\item{rnd}{Vector of quantiles}

\item{rate}{The rate parameter

Note taht the conditional quantile for an exponential is independent of time due to constant hazard}
}
\value{
Estimate(s) from the conditional exponential distribution based on given parameters
}
\description{
Conditional quantile function for exponential distribution
}
\examples{
qcond_exp(rnd = 0.5,rate = 3)
}
