% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qcond_llogis}
\alias{qcond_llogis}
\title{Conditional quantile function for loglogistic distribution}
\usage{
qcond_llogis(rnd, shape, scale, lower_bound = as.numeric(c(0)))
}
\arguments{
\item{rnd}{Vector of quantiles}

\item{shape}{The shape parameter}

\item{scale}{The scale parameter}

\item{lower_bound}{The lower bound to be used (current time)}
}
\value{
Estimate(s) from the conditional loglogistic distribution based on given parameters
}
\description{
Conditional quantile function for loglogistic distribution
}
\examples{
qcond_llogis(rnd = 0.5,shape = 1,scale = 1,lower_bound = 1)
}
