% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{qcond_weibull}
\alias{qcond_weibull}
\title{Conditional quantile function for weibull distribution}
\usage{
qcond_weibull(rnd = 0.5, shape, scale, lower_bound = 0)
}
\arguments{
\item{rnd}{Vector of quantiles}

\item{shape}{The shape parameter as in R stats package weibull}

\item{scale}{The scale parameter as in R stats package weibull}

\item{lower_bound}{The lower bound to be used (current time)}
}
\value{
Estimate(s) from the conditional weibull distribution based on given parameters
}
\description{
Conditional quantile function for weibull distribution
}
\examples{
qcond_weibull(rnd = 0.5,shape = 3,scale = 66.66,lower_bound = 50)
}
