% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{qgamma_mse}
\alias{qgamma_mse}
\title{Use quantiles from a gamma distribution based on mean and se}
\usage{
qgamma_mse(q = 1, mean_v, se, seed = NULL)
}
\arguments{
\item{q}{Quantile to draw}

\item{mean_v}{A vector of the mean values}

\item{se}{A vector of the standard errors of the means}

\item{seed}{An integer which will be used to set the seed for this draw.}
}
\value{
A single estimate from the gamma distribution based on given parameters
}
\description{
Use quantiles from a gamma distribution based on mean and se
}
\examples{
qgamma_mse(q=0.5,mean_v=0.8,se=0.2)

}
