% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_f.R
\name{qresgompertz}
\alias{qresgompertz}
\title{Quantile function for restricted Gompertz distribution (lower bound only)}
\usage{
qresgompertz(rnd = 0.5, shape, rate, lower_bound)
}
\arguments{
\item{rnd}{Vector of quantiles}

\item{shape}{The shape parameter of the Gompertz distribution, defined as in the coef() output on a flexsurvreg object}

\item{rate}{The rate parameter of the Gompertz distribution, defined as in the coef() output on a flexsurvreg object}

\item{lower_bound}{The lower bound of the restricted distribution}
}
\value{
Estimate(s) from the restricted Gompertz distribution based on given parameters
}
\description{
Quantile function for restricted Gompertz distribution (lower bound only)
}
\examples{
qresgompertz(rnd=0.5,shape=0.05,rate=0.01,lower_bound = 50)
}
