% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{queue_create}
\alias{queue_create}
\title{Create a New Event Queue}
\usage{
queue_create(priority_order)
}
\arguments{
\item{priority_order}{A character vector of event names sorted by decreasing importance.}
}
\value{
An external pointer to the new event queue.
}
\description{
Initializes a new event queue with the specified priority order of event names.
}
