% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{queue_size}
\alias{queue_size}
\title{Get the Size of the Event Queue}
\usage{
queue_size(ptr, exclude_inf = FALSE)
}
\arguments{
\item{ptr}{The event queue pointer. Defaults to \code{cur_evtlist}.}

\item{exclude_inf}{Logical, whether to exclude events with Inf time. Default is FALSE.}
}
\value{
An integer indicating the number of events in the queue.
}
\description{
Get the Size of the Event Queue
}
