% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{revert_list}
\alias{revert_list}
\title{Reverts list}
\usage{
revert_list(ls)
}
\arguments{
\item{ls}{List}
}
\value{
Reverted list
}
\description{
Reverts list
}
\examples{
\dontrun{
revert_list(ls=list(a=list(ab=1,ac=2),b=list(ab=2,ac=3)))
}

}
