% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{shared_incr}
\alias{shared_incr}
\title{Increment a shared counter}
\usage{
shared_incr(obj, delta = 1)
}
\arguments{
\item{obj}{A \code{shared_input} object (constrained / shared mode).}

\item{delta}{Numeric. Amount to add (default \code{1}).}
}
\value{
The new value after incrementing.
}
\description{
Adds \code{delta} to the value of a \code{shared_input} object and returns the new
value.
}
