% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary_f.R
\name{summary_results_sens}
\alias{summary_results_sens}
\title{Summary of sensitivity outputs for a treatment}
\usage{
summary_results_sens(out = results, arm = NULL, wtp = 50000)
}
\arguments{
\item{out}{The list object returned by \code{run_sim()}}

\item{arm}{The reference treatment for calculation of incremental outcomes}

\item{wtp}{Willingness to pay to have INMB}
}
\value{
A data frame with each sensitivity output per arm
}
\description{
Summary of sensitivity outputs for a treatment
}
\examples{
\dontrun{
summary_results_sens(results, arm="int")
}
}
