% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tte.df}
\alias{tte.df}
\title{Example TTE IPD data}
\format{
\subsection{\code{tte.df}}{

A data frame with 1000 rows and 8 columns:
\describe{
\item{USUBJID}{Patient ID}
\item{ARMCD, ARM}{Arm code and variables}
\item{PARAMCD, PARAM}{Parameter}
\item{AVAL, AVALCD}{Values of interest}
\item{CNSR}{Censored observation?}
}
}
}
\source{
Simulated through FlexsurvPlus package using sim_adtte(seed = 821, rho = 0, beta_1a = log(0.6), beta_1b = log(0.6), beta_pd = log(0.2))
}
\usage{
tte.df
}
\description{
An example of TTE IPD data for the example_ipd file
}
\keyword{datasets}
