% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{disc_cycle_v}
\alias{disc_cycle_v}
\title{Cycle discounting for vectors}
\usage{
disc_cycle_v(
  lcldr = 0.035,
  lclprvtime = 0,
  cyclelength,
  lclcurtime,
  lclval,
  starttime = 0
)
}
\arguments{
\item{lcldr}{The discount rate}

\item{lclprvtime}{The time of the previous event in the simulation}

\item{cyclelength}{The cycle length}

\item{lclcurtime}{The time of the current event in the simulation}

\item{lclval}{The  value to be discounted}

\item{starttime}{The start time for accrual of cycle costs (if not 0)}
}
\value{
Double based on cycle discounting
}
\description{
Cycle discounting for vectors
}
\examples{
disc_cycle_v(lcldr=0.035, lclprvtime=0, cyclelength=1/12, lclcurtime=2, lclval=500,starttime=0)


}
