% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{extract_from_reactions}
\alias{extract_from_reactions}
\title{Extract all items and events and their interactions from the event reactions list}
\usage{
extract_from_reactions(reactions)
}
\arguments{
\item{reactions}{list generated through \code{add_reactevt}}
}
\value{
A data.frame with the relevant item/event, the event where it's assigned,
and whether it's contained within a conditional statement
}
\description{
Extract all items and events and their interactions from the event reactions list
}
\examples{
a <- add_reactevt(name_evt="example",
   input={
      modify_item(list(w=5))
   })
   
 extract_from_reactions(a)


}
