% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{has_event}
\alias{has_event}
\title{Check if a Patient Has a Specific Event}
\usage{
has_event(event_name, ptr, patient_id, exclude_inf = FALSE)
}
\arguments{
\item{event_name}{Character string, the name of the event.}

\item{ptr}{The event queue pointer. Defaults to \code{cur_evtlist}.}

\item{patient_id}{The patient ID. Defaults to \code{i}.}

\item{exclude_inf}{Logical, whether to exclude events with Inf time. Default is FALSE.}
}
\value{
Logical, TRUE if the event exists for the patient (optionally excluding Inf), FALSE otherwise.
}
\description{
Check if a Patient Has a Specific Event
}
