% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{next_event}
\alias{next_event}
\title{Get the Next Event(s) in the Queue}
\usage{
next_event(n = 1, ptr)
}
\arguments{
\item{n}{Number of events to retrieve. Default is 1.}

\item{ptr}{The event queue pointer. Defaults to \code{cur_evtlist}.}
}
\value{
A list of events, each with \code{patient_id}, \code{event_name}, and \code{time}.
}
\description{
Retrieves the next \code{n} events (without removing them).
}
