% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{next_event_pt}
\alias{next_event_pt}
\title{Get the Next Event(s) in the Queue for a specific patient}
\usage{
next_event_pt(n = 1, ptr, patient_id)
}
\arguments{
\item{n}{Number of events to retrieve. Default is 1.}

\item{ptr}{The event queue pointer. Defaults to \code{cur_evtlist}.}

\item{patient_id}{The patient ID. Defaults to \code{i}.}
}
\value{
A list of events, each with \code{patient_id}, \code{event_name}, and \code{time}.
}
\description{
Retrieves the next \code{n} events (without removing them).
}
