% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{pop_and_return_event}
\alias{pop_and_return_event}
\title{Pop and Return the Next Event}
\usage{
pop_and_return_event(ptr)
}
\arguments{
\item{ptr}{The event queue pointer. Defaults to \code{cur_evtlist}.}
}
\value{
A named list with \code{patient_id}, \code{event_name}, and \code{time}.
}
\description{
Removes the next event from the queue and returns its details. Not needed by user.
}
