% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{queue_empty}
\alias{queue_empty}
\title{Check if the Event Queue is Empty}
\usage{
queue_empty(ptr, exclude_inf = FALSE)
}
\arguments{
\item{ptr}{The event queue pointer. Defaults to \code{cur_evtlist}.}

\item{exclude_inf}{Logical, whether to exclude events with Inf time. Default is FALSE.}
}
\value{
Logical, TRUE if the queue is empty, FALSE otherwise.
}
\description{
Check if the Event Queue is Empty
}
