% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_f.R
\name{remove_event}
\alias{remove_event}
\title{Remove Events for a Patient}
\usage{
remove_event(events, ptr, patient_id)
}
\arguments{
\item{events}{A character vector of event names to remove. It can also handle lists instead of named vectors (at a small computational cost).}

\item{ptr}{The event queue pointer. Defaults to \code{cur_evtlist}.}

\item{patient_id}{The patient ID. Defaults to \code{i}.}
}
\value{
NULL (invisible). Modifies the queue in-place.
}
\description{
Removes one or more events from the queue for the given patient.
}
