% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary_f.R
\name{summary_results_sim}
\alias{summary_results_sim}
\title{Summary of PSA outputs for a treatment}
\usage{
summary_results_sim(out = results[[1]], arm = NULL, wtp = 50000)
}
\arguments{
\item{out}{The output_sim data frame from the list object returned by \code{run_sim()}}

\item{arm}{The reference treatment for calculation of incremental outcomes}

\item{wtp}{Willingness to pay to have INMB}
}
\value{
A data frame with mean and 95\% CI of absolute costs, LYs, QALYs, ICER and ICUR for each intervention from the PSA samples
}
\description{
Summary of PSA outputs for a treatment
}
\examples{
\dontrun{
summary_results_sim(results[[1]], arm="int")
}
}
